Fixes #621750 - Make Help Icons smaller
- Reduces the size of the help icons that appear next to labels to
13x13px.

- Replaces the '?' character with an image, to ensure proper positioning
and size.

- Shows the hover-tip if the user hovers over the title or the icon, not
just the icon (ie the whole content of the tip() macro).

- Attempts to uniformize most calls to the tip(name, tip) macro to
include the label title as the first argument, and allow the tip to how
on top of the title too instead of just the '?' icon. Most calls to
tip() already were already doing that, but a few were displaying the
title separately, passing None as a first argument.